Add token authentication
This sets up basic token authentication.  All of the logic is currently
in the notes controller and should most likely be moved some place
common where multiple controllers could use it.

We still do not have any routes that allows a user to sign in and
recieve their authentication token.